// src/components/search/types.ts
export type PagespeedDataType = {
        timestamp: Date;
        url: string;
        cumulative_layout_shift: string;
        first_contentful_paint: string;
        server_response_time: string;
        total_blocking_time: string;
        speed_index: string;
        time_to_interactive: string;
        first_meaningful_paint: string;
    
}